requirements.txt
robotframework
robotframework-faker
robotframework-pythonlibcore
sqlalchemy
psycopg2-binary
factory-boy